pyrxing/requirements.txt
pytest==8.3.5
pillow==11.1.0
